backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
pydantic==2.0.3